App with Flask/requirements.txt
flask
torch --index-url https://download.pytorch.org/whl/cpu
torchvision --index-url https://download.pytorch.org/whl/cpu
segmentation-models-pytorch
rasterio
numpy
Pillow